move git.ts to not just be a typescript definition
